Revert "reorganized app folder structure adding /user and /navigation folders…"
