+++ src/app/app.js
@@ -2,9 +2,14 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
 import { Provider } from "react-redux"
+import { BrowserRouter as Router, Route , IndexRoute } from 'react-router-dom'
+
+import Main from './containers/Main.js';
+import Welcome from './containers/welcome.js';
+
 // import routes from './react-routes'
 import configureStore from './store';
-import AppRoutes from './navigation/AppRoutes.js';
+import AppRoutes from './react-routes';
 
 const store = configureStore();
