+++ src/app/containers/SignUpPageContainer.js
@@ -1,7 +1,7 @@
-import LogInPage from './LogInPage.js';
+import SignUpPage from '../components/SignUpPage.js';
 import { connect } from "react-redux";
 import { bindActionCreators } from 'redux';
-import { loginUser } from '../redux/userActions.js'
+import { loginUser } from '../actions/userActions.js'
 
 function mapStateToProps(state) {
   return {
@@ -14,4 +14,4 @@ function mapDispatchToProps(dispatch){
   return bindActionCreators({ loginUser }, dispatch);
 }
 
-export default connect(mapStateToProps, mapDispatchToProps)(LogInPage);
+export default connect(mapStateToProps, mapDispatchToProps)(SignUpPage);
