+++ src/app/reducers/index.js
@@ -1,12 +1,12 @@
 //Here all we need is the combineReducers function to combine our reducers so we can export them to our store
 import { combineReducers } from 'redux';
-import { reducer as UserReducer } from '../users';
+import UserReducer from './userReducer.js';
 import { reducer as burgerMenu } from 'redux-burger-menu';
 import { reducer as formReducer } from 'redux-form';
 
 const rootReducer = combineReducers({
   user: UserReducer,
-  burgerMenu: burgerMenu,
+  burgerMenu:   burgerMenu,
   form: formReducer,
 });
