--- src/app/users/signup/SignUpPageContainer.js
@@ -1,17 +1,0 @@
-import SignUpPage from './SignUpPage.js';
-import { connect } from "react-redux";
-import { bindActionCreators } from 'redux';
-import { loginUser } from '../redux/userActions.js'
-
-function mapStateToProps(state) {
-  return {
-    user: state.user,
-    isAuth: state.user.isAuth,
-  };
-}
-
-function mapDispatchToProps(dispatch){
-  return bindActionCreators({ loginUser }, dispatch);
-}
-
-export default connect(mapStateToProps, mapDispatchToProps)(SignUpPage);
